internal/featuretests: move upstream ca validation tests
Updates #1436

Move upstream CA validation tests to internal/featuretests.
